#7 fix: db migration is not executed when using legacy backend
