+++ plugins/infrawallet-backend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@electrolux-oss/plugin-infrawallet-backend",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "backstage": {
     "role": "backend-plugin"
   },
